Combine LC and GC into one platform
