+++ application/src/main/resources/config/application-development-lc.properties
@@ -1,0 +1,17 @@
+# Settings specific for the "development" profile
+# Overrides default values
+logging.file=${user.home}/.config/de.hswt.fi/logs/fi-development.log
+
+spring.datasource.user.driverClassName=org.postgresql.Driver
+spring.datasource.user.url=jdbc:postgresql://localhost:5432/henry-user
+spring.datasource.user.username=postgres
+spring.datasource.user.password=postgres
+spring.datasource.user.test-on-borrow=true
+spring.datasource.user.validation-query=SELECT 1
+
+spring.datasource.compound.driverClassName=org.postgresql.Driver
+spring.datasource.compound.url=jdbc:postgresql://localhost:5432/henry-data
+spring.datasource.compound.username=postgres
+spring.datasource.compound.password=postgres
+spring.datasource.compound.test-on-borrow=true
+spring.datasource.compound.validation-query=SELECT 1
