+++ application/src/main/resources/config/application-gc.properties
@@ -1,0 +1,1 @@
+application-test.properties=FOR-IDENT GC
