+++ application/src/main/resources/config/application-lc.properties
