+++ application/src/main/resources/config/application.properties
@@ -11,7 +11,7 @@ de.hswt.fi.security.logout-url=/logout
 de.hswt.fi.security.password-length=10
 de.hswt.fi.security.registration-enabled=true
 
-de.hswt.fi.ui.header.caption=DUFTSTOFF-IDENT
+de.hswt.fi.ui.header.caption=FOR-IDENT
 de.hswt.fi.ui.header.collapsed-caption=FI
 de.hswt.fi.ui.header.visible=false
 de.hswt.fi.ui.links.reach=http://echa.europa.eu/web/guest/information-on-chemicals/registered-substances
@@ -35,3 +35,5 @@ html-home-english=home_en.html
 html-home-german=home_de.html
 html-help-processing-english=processing_help_en.html
 html-help-processing-german=processing_help_de.html
+
+logging.level.org.springframework.beans = TRACE
