+++ common/src/main/java/de/hswt/fi/common/spring/SpringProfileUtil.java
@@ -28,7 +28,7 @@ public class SpringProfileUtil {
 	}
 
 	public boolean isDevelopmentProfile() {
-		return activeProfiles.contains(Profiles.DEVELOPMENT);
+		return activeProfiles.contains(Profiles.DEVELOPMENT_GC) || activeProfiles.contains(Profiles.DEVELOPMENT_LC);
 	}
 
 	public List<String> getActiveProfiles() {
