+++ fileimport.service.excel.formats/src/main/resources/config/rti_calibration_config_logd.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<file format="97-2007">
+    <sheets>
+        <sheet id="de.hswt.fi.rti.excel.calibration.sheet" identifierRow="0" contentRow="3" optional="false">
+        	<groups>
+        		<group id = "G1" requiredColumns = "all">
+		            <columns>
+        	        	<column id ="de.hswt.fi.rti.excel.calibration.substancename" name="substance name" />
+            	    	<column id ="de.hswt.fi.rti.excel.calibration.signal" name="logD" />
+                		<column id ="de.hswt.fi.rti.excel.calibration.rt" name="rt" multiple="true" />
+		            </columns>
+				</group>
+			</groups>
+        </sheet>
+    </sheets>
+</file>
