+++ fileimport.service.excel.formats/src/test/java/de/hswt/fi/fileimport/service/excel/formats/ExcelImportCalibrationTest.java
@@ -30,7 +30,7 @@ public class ExcelImportCalibrationTest {
 	@Before
 	public void setupReferenceData() {
 
-		calibrationReader = new CalibrationExcelReader();
+		calibrationReader = new CalibrationExcelReaderGC();
 
 		data = new ArrayList<>();
