+++ fileimport.service.excel.formats/src/test/java/de/hswt/fi/fileimport/service/excel/formats/ExcelImportSciexMasterViewMSMSTest.java
@@ -18,7 +18,7 @@ public class ExcelImportSciexMasterViewMSMSTest extends AbstractImportTest {
 	public FileBasedFeatureContentImporter getFileImporter() {
 		return new ExcelFeatureContentImporter(
 				Collections.singletonList(new SciexMSMSExcelReader()),
-				Collections.singletonList(new CalibrationExcelReader()));
+				Collections.singletonList(new CalibrationExcelReaderGC()));
 	}
 
 	@Override
