+++ fileimport.service.excel.formats/src/test/java/de/hswt/fi/fileimport/service/excel/formats/ExcelImportThermoCiecveTest.java
@@ -18,7 +18,7 @@ public class ExcelImportThermoCiecveTest extends AbstractImportTest {
 	public FileBasedFeatureContentImporter getFileImporter() {
 		return new ExcelFeatureContentImporter(
 				Collections.singletonList(new ThermoCiecveExcelReader()),
-				Collections.singletonList(new CalibrationExcelReader()));
+				Collections.singletonList(new CalibrationExcelReaderGC()));
 	}
 
 	@Override
