+++ model/src/main/resources/de/hswt/fi/model/i18n/messages_en.properties
@@ -9,5 +9,6 @@ FEATURE_PRECURSOR_MASS=Precursor Mass
 FEATURE_FORMULA=Formula
 FEATURE_RT=Retention Time
 FEATURE_RTI=Retention Time Index
+FEATURE_RT_SIGNAL=Retention Time Signal
 FEATURE_FORMULA_DERIVED_MASS=Mass (formula)
 FEATURE_NEUTRAL_MASS=Masse (neutral)
