+++ security.service.vaadin/src/main/java/de/hswt/fi/security/service/vaadin/configuration/RestrictedAccessConfiguration.java
@@ -1,12 +1,10 @@
 package de.hswt.fi.security.service.vaadin.configuration;
 
 import de.hswt.fi.application.properties.ApplicationProperties;
-import de.hswt.fi.common.spring.Profiles;
 import de.hswt.fi.security.service.vaadin.CustomUserDetailsService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
@@ -32,7 +30,6 @@ import org.vaadin.spring.security.shared.VaadinSessionClosingLogoutHandler;
  *
  */
 
-@Profile(value = {Profiles.PRODUCTION, Profiles.DEVELOPMENT})
 @Configuration
 @EnableWebSecurity
 @EnableGlobalMethodSecurity(securedEnabled = true, prePostEnabled = true, proxyTargetClass = true)
